link READMEto 2015-Code demo and update GPL info
Updated the GPL info stated in README.md to specify that we are fine
with the use of future versions of the GPL liscense (@zachohara
@pranavmathur @thunderbots5604/contributors).

Also made the reference to 2015-Code link to the repository. This makes
it easier for teams just starting with Github to find their way to it.

Feel free to change the liscense back to only GPLv3 if we want to.