Rename 'DriveMotorSet' to 'MotorSet' and move it to the 'hardware' package
The concept of a motor set could apply to more purposes than just driving. The example that comes to my mind is the lift system that we used on our 2013-14 robot. The lift was controlled by two different physical motors, but any variation in speed would cause serious problems for the gear system.