Revert "Line Wrap Comments to <90 characters"
This reverts commit 443718136ffe4625e0e075d09d77ee4d8e688caf.

@PranavMathur
The line wrap fix should be in its own commit, and it should include all the files in the repository. Additionally, it should be made to master, not to joystick efficiency. Eclipse has a great auto-format feature that we can use, and that will fix more things than just line wrap.